refactor: Reorganize user functions and tests to be in their own files
